// src/app/core/models/card.ts
export interface Card {
    id: number; // 1
    fullName: string; // họ và tên: Phạm Hoàng Dủ
    religiousName: string; // pháp danh: Đồng Khả Nhân
    registerCode: string; // mã đăng ký: HD1-001
    organzationStuctureName: string; // tên CTN: CTN-HCM
    avatarPath: string; // link hình thẻ: https://ctnpq.com/avatar/template.jpg
    departmentCode: string; // mã ban: HD1
}